* configure.in: AC_REPLACE_FUNCS(fgetpwent);
	* include/mailutils/parse822.h: Add a prototype for
	parse822_skip_nl().

	* lib/Makefile.am: add fgetpwent.c
	* lib/fgetpwent.c: New file.

	* mail/from.c (mail_from): Use the cover functions
	attribute_is_xxxx() instead of getting the flag.

	* mailbox/attribute.c: Cleanup the mess.
	attribute_set_{read,delete,seen,flagged,answered}.
	attribute_unset_{read,delete,seen,flagged,answered}.
	attribute_is_{read,delete,seen,flagged,answered}
	should be no more then cover functions calling
	attribute_set_flags(), attribute_get_flags() and
	attribute_unset_flags().

	* mailbox/folder_imap.c: MU_ATTRIBUTE_READ, is not
	part of IMAP protocol use \\Seen.
	* mailbox/mbx_imap.c: Remove the hack trying to reconnect,
	it does not work.

	* mailbox/mbx_pop.c(pop_get_messages): Set the new
	functions for the attribute.
	(pop_get_attribute): New functions.
	(pop_set_attribute): New functions.
	(pop_unset_attribute): New functions.
	(pop_attr_flags): Removed.

	* mutil.c (mu_hex2ul): GNU coding std.
	(parse822_time): Do not use tzname as a variable name tzname
	is global variable in libc, better clear away.
	(parse822_date_time): Likewised.

	* pop3d/Makefile.am: add virtual.c
	* pop3d/pop3d.c (main): register getpwnam_ip_virtual,
	getpwnam_host_virtual
	* pop3d/pop3d.h: declare,
	getpwnam_ip_virtual(), getpwnam_host_virtual().
	* pop3d/virtual.c: New file.

+++ mailbox/mutil.c
@@ -35,7 +35,8 @@
 
 /* convert a sequence of hex characters into an integer */
 
-unsigned long mu_hex2ul(char hex)
+unsigned long
+mu_hex2ul (char hex)
 {
   if (hex >= '0' && hex <= '9')
     return hex - '0';
@@ -49,21 +50,22 @@ unsigned long mu_hex2ul(char hex)
   return -1;
 }
 
-size_t mu_hexstr2ul(unsigned long* ul, const char* hex, size_t len)
+size_t
+mu_hexstr2ul (unsigned long *ul, const char *hex, size_t len)
 {
   size_t r;
 
   *ul = 0;
 
   for (r = 0; r < len; r++)
-  {
-    unsigned long v = mu_hex2ul(hex[r]);
+    {
+      unsigned long v = mu_hex2ul (hex[r]);
 
-    if(v == (unsigned long)-1)
-      return r;
+      if (v == (unsigned long)-1)
+	return r;
 
-    *ul = *ul * 16 + v;
-  }
+      *ul = *ul * 16 + v;
+    }
   return r;
 }
 
@@ -83,18 +85,18 @@ mu_tm2time (struct tm *timeptr, mu_timezone* tz)
 {
   int offset = tz ? tz->utc_offset : 0;
 
-  return mktime(timeptr) + mu_utc_offset() - offset;
+  return mktime (timeptr) + mu_utc_offset () - offset;
 }
 
 /* Convert time 0 at UTC to our localtime, that tells us the offset
    of our current timezone from UTC. */
 time_t
-mu_utc_offset(void)
+mu_utc_offset (void)
 {
   time_t t = 0;
-  struct tm* tm = gmtime(&t);
+  struct tm* tm = gmtime (&t);
 
-  return - mktime(tm);
+  return - mktime (tm);
 }
 
 static const char *months[] =
@@ -243,7 +245,6 @@ mu_parse_ctime_date_time (const char **p, struct tm *tm, mu_timezone * tz)
   return 0;
 }
 
-
 char *
 mu_get_homedir (void)
 {
@@ -252,7 +253,7 @@ mu_get_homedir (void)
     {
       struct passwd *pwd;
 
-      pwd = getpwuid(getuid());
+      pwd = getpwuid (getuid ());
       if (!pwd)
 	return NULL;
       homedir = pwd->pw_dir;
@@ -369,7 +370,7 @@ getpwnam_virtual (const char *u)
 {
   struct passwd *pw = NULL;
   FILE *pfile;
-  int i = 0, len = strlen (u), delim = 0;
+  size_t i = 0, len = strlen (u), delim = 0;
   char *filename;
 
   mu_virtual_domain = 0;
